+++ client/client.go
@@ -325,16 +325,14 @@ type PrmDial struct {
 //
 // Format of the URI:
 //
-//	[scheme://]host:port
+//	[scheme://]host[:port]
 //
 // Supported schemes:
 //
 //	grpc
 //	grpcs
 //
-// If scheme is omitted, port 443 implies grpcs (TLS), any other port implies
-// grpc (no TLS). If grpcs scheme is specified without a port, port 443 is
-// used by default.
+// If grpcs scheme is specified without a port, port 443 is used by default.
 //
 // See also SetTLSConfig.
 func (x *PrmDial) SetServerURI(endpoint string) {